+++ project-base/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- `JavascriptCompilerService` can now compile javascript from more than one source directory (@MattCzerner)
+    - the compiler supports subdirectory `common` in addition to `admin` and  `frontend`
 
 ## 2.0.0-beta.21.0 - 2018-03-05
 ### Added